# Performance optimizations for large product stacks