A series of performance optimizations for large installations (thousands of products), identified through profiling.

Together with related optimizations in sconsUtils, speeds up the product builds on lsst-dev by 1.5x-2.5x, depending on the product. E.g.:

Before:
```
ndarray: master-g7408a83a3a+4 ................................................ok (138.8 sec).
daf_persistence: 10.0+23 ......................................................ok (124.6 sec).
```
After:
```
ndarray: master-g7408a83a3a+5 .......................ok (87.5 sec).
daf_persistence: 10.0+23 ....................ok (55.2 sec).
```