@mjuric 2015-03-27
@RobertLuptonTheGood @ktlim I pushed the new version addressing review comments, take another look when you get a chance.