@ktlim 2015-03-27
Aside from the minor issues above, OK to merge from me.